// lib/api/types.ts
export interface ProductDto {
  id: number
  user_id?: number
  nomi?: string
  kategoriya?: string
  birlik?: string
  olish_narxi?: number
  sotish_narxi?: number
  min_sotish_narxi?: number
  qoldiq?: number
  min_qoldiq?: number
  yaratilgan?: string
  // Allow optional frontend-expected aliases
  sku?: string
  narx?: number
  ombor?: number
  min_ombor?: number
  tavsif?: string
  holat?: string
}